+++ lib/montague.rb
@@ -1,4 +1,4 @@
-require 'montague/api'
+require 'montague/api/api'
 require 'montague/xml_extractor/xml_extractor'
 
 # Wrapper for the SHERPA RoMEO API