+++ lib/montague/api/api.rb
@@ -1,0 +1,10 @@
+require 'http'
+require 'montague/api/base'
+require 'montague/api/journal'
+
+module Montague
+  # Requests for a resource
+  #
+  module API
+  end
+end