+++ lib/montague/xml_extractor/xml_extractor.rb
@@ -1,6 +1,7 @@
 require 'nokogiri'
 require 'montague/xml_extractor/base'
 require 'montague/xml_extractor/journal'
+require 'montague/xml_extractor/journals'
 
 module Montague
   # An XMLExtractor manages the extraction of metadata from XML into Ruby