+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.1.0 - 2018-01-23
+## 0.1.0 - 2018-01-24
 ### Added
-- Find by ISSN.
+- For SHERPA/RoMEO API v2.9.
+- Find journal by ISSN, title.
+- Find journals by partial title.